+++ src/__tests__/helpers.js
@@ -2,16 +2,16 @@ import { exec as execCallback } from 'node:child_process'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { promisify } from 'node:util'
+import { chromium } from '@playwright/test'
 
 const exec = promisify(execCallback)
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 const extensionPath = path.join(__dirname, '../../dist')
 
-export const launchPuppeteerWithExtension = async function (puppeteer) {
+export const launchPlaywrightWithExtension = function () {
   const options = {
     headless: false,
     ignoreHTTPSErrors: true,
-    devtools: true,
     args: [
       `--disable-extensions-except=${extensionPath}`,
       `--load-extension=${extensionPath}`,
@@ -20,13 +20,11 @@ export const launchPuppeteerWithExtension = async function (puppeteer) {
     ],
   }
 
-  if (process.env.PUPPETEER_EXEC_PATH) {
-    options.executablePath = process.env.PUPPETEER_EXEC_PATH
-  } else {
-    options.executablePath = await puppeteer.executablePath()
+  if (process.env.PLAYWRIGHT_CHROMIUM_EXECUTABLE_PATH) {
+    options.executablePath = process.env.PLAYWRIGHT_CHROMIUM_EXECUTABLE_PATH
   }
 
-  return puppeteer.launch(options)
+  return chromium.launchPersistentContext('', options)
 }
 
 export const runBuild = function () {