+++ src/modules/code-generator/constants.js
@@ -1,10 +1,6 @@
 export const headlessActions = {
   GOTO: 'GOTO',
   VIEWPORT: 'VIEWPORT',
-  WAITFORSELECTOR: 'WAITFORSELECTOR',
-  NAVIGATION: 'NAVIGATION',
-  NAVIGATION_PROMISE: 'NAVIGATION_PROMISE',
-  FRAME_SET: 'FRAME_SET',
   SCREENSHOT: 'SCREENSHOT',
 }
 
@@ -17,9 +13,4 @@ export const eventsToRecord = {
   SUBMIT: 'submit',
   LOAD: 'load',
   UNLOAD: 'unload',
-}
-
-export const headlessTypes = {
-  PUPPETEER: 'puppeteer',
-  PLAYWRIGHT: 'playwright',
 }